organised the e2e tests into a few more page objects. need to delete fake test data from db, this can cause errors in this e2e tests.
